1. Rename sniff. 2. Adjust the logic per https://github.com/WordPress/gutenberg/issues/52769#issuecomment-1669540435.
